refactor(server): change to vertical slice architecture.
Slicing based on feature on src folder.
