+++ server/.env
@@ -1,0 +1,1 @@
+PORT=5000
